Fix most of code review issues
